+++ docs/conf.py
@@ -12,7 +12,6 @@ import os
 import re
 import subprocess
 import sys
-import textwrap
 import time
 from functools import lru_cache, partial
 from typing import Any, Callable, Dict, Iterable, Iterator, List, Tuple
@@ -31,7 +30,6 @@ if kitty_src not in sys.path:
 from kitty.conf.types import Definition, expand_opt_references  # noqa
 from kitty.constants import str_version, website_url # noqa
 from kitty.fast_data_types import Shlex, TEXT_SIZE_CODE  # noqa
-from kittens.panel.main import default_quake_cmdline  # noqa
 
 # config {{{
 # -- Project information -----------------------------------------------------
@@ -122,9 +120,6 @@ string_replacements = {
     '_kitty_install_cmd': 'curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin',
     '_build_go_version': go_version('../go.mod'),
     '_text_size_code': str(TEXT_SIZE_CODE),
-    '_default_quake_cmdline': textwrap.fill(
-        default_quake_cmdline, break_on_hyphens=False, break_long_words=False, initial_indent=' ' * 4, subsequent_indent=' ' * 8, width=77,
-    ).replace('\n', ' \\\n'),
 }